resolved Windows file permission error

+++ metaflow/client/filecache.py
@@ -7,7 +7,12 @@ import time
 from tempfile import NamedTemporaryFile
 from hashlib import sha1
 
-from urllib.parse import urlparse
+try:
+    # Python 3
+    from urllib.parse import urlparse
+except ImportError:
+    # Python 2
+    from urlparse import urlparse
 
 from metaflow.datastore import FlowDataStore
 from metaflow.datastore.content_addressed_store import BlobCache
@@ -113,7 +118,8 @@ class FileCache(object):
 
         log = task_ds.load_log_legacy(logtype, attempt_override=attempt)
         # Store this in the file cache as well
-        self.create_file(path, log)
+        if log is not None:
+            self.create_file(path, log)
         return log
 
     def get_legacy_log_size(
@@ -263,16 +269,22 @@ class FileCache(object):
         dirname = os.path.dirname(path)
         try:
             FileCache._makedirs(dirname)
-        except:  # noqa E722
+        except Exception:  # noqa E722
             raise FileCacheException("Could not create directory: %s" % dirname)
         tmpfile = NamedTemporaryFile(dir=dirname, prefix="dlobj", delete=False)
         # Now write out the file
         try:
             tmpfile.write(value)
             tmpfile.flush()
-            os.rename(tmpfile.name, path)
-        except:  # noqa E722
-            os.unlink(tmpfile.name)
+            os.fsync(tmpfile.fileno())
+            tmpfile.close()
+
+            os.replace(tmpfile.name, path)
+
+        except Exception:  # noqa E722
+            tmpfile.close()
+            if os.path.exists(tmpfile.name):
+                os.unlink(tmpfile.name)
             raise
         size = os.path.getsize(path)
         self._total += size
@@ -305,8 +317,8 @@ class FileCache(object):
                         sha, ext = os.path.splitext(obj)
                         if ext in ["cached", "blob"]:
                             path = os.path.join(root, obj)
-                            objects.insert(
-                                0, (os.path.getctime(path), os.path.getsize(path), path)
+                            objects.append(
+                                (os.path.getctime(path), os.path.getsize(path), path)
                             )
 
         self._total = sum(size for _, size, _ in objects)
